+++ pkg/packy/packer.go
@@ -1,14 +1,16 @@
 package packy
 
+// Packer packs blocks
 type Packer struct {
 	root *Node
 }
 
-func NewPacker(w, h int) *Packer {
+// NewPacker has packing abilities
+func NewPacker(w, h float64) *Packer {
 	return &Packer{root: &Node{W: w, H: h}}
 }
 
-func (pk *Packer) findNode(root *Node, w, h int) *Node {
+func (pk *Packer) findNode(root *Node, w, h float64) *Node {
 	var node *Node
 	if root.used {
 		node = pk.findNode(root.right, w, h)
@@ -22,7 +24,7 @@ func (pk *Packer) findNode(root *Node, w, h int) *Node {
 	return node
 }
 
-func (pk *Packer) splitNode(node *Node, w, h int) *Node {
+func (pk *Packer) splitNode(node *Node, w, h float64) *Node {
 	node.used = true
 	node.down = &Node{X: node.X, Y: node.Y + h, W: node.W, H: node.H - h}
 	node.right = &Node{X: node.X + w, Y: node.Y, W: node.W - w, H: h}
@@ -30,6 +32,7 @@ func (pk *Packer) splitNode(node *Node, w, h int) *Node {
 	return node
 }
 
+// Fit try to fit block
 func (pk *Packer) Fit(blocks []*Node) {
 	var (
 		n     int
