+++ pkg/packy/type.go
@@ -1,7 +1,8 @@
 package packy
 
+// Node is a rectangle like repesenting a block
 type Node struct {
-	X, Y, W, H       int
+	X, Y, W, H       float64
 	used             bool
 	right, down, Fit *Node
 }
